greenwood_history_store/greenwood_history/apps/entities/serializers.py
from rest_framework import serializers

from .models import Entity


class EntitySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Entity
        fields = ["name", "meta"]
