greenwood_history_store/pyproject.toml
[tool.poetry]
name = "greenwood_history_store"
version = "0.1.0"
description = "Data store of events/plots/people in greenwood"
authors = ["Greenwood mapping team"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
django-configurations = {version = "^2.3.2", extras = ["cache", "database", "email", "search"]}
rich = "^12.5.1"
pydantic = "^1.9.2"
django-pydantic-field = "^0.2.3"
django-polymorphic = "^3.1.0"
nameparser = "^1.1.1"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
bonobo = "^0.6.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
isort = "^5.10.1"
pdbr = "^0.7.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '/(\.eggs|\.git|\.mypy_cache|\.pytest_cache|\.tox|venv|\.venv|build|dist|migrations|htmlcov)/'
include = '\.pyi?$'
line-length = 99
target-version = ['py38']

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 99
indent = "    "
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
skip_glob = "*/migrations/*"
# known_local_folder = ["greenwood/"]

[tool.pdbr]
store_history=".pdbr_history"
style = "yellow"
use_traceback = true
theme = "friendly"